core: add new queue
It is basically the same as the old queue, but does not keep data
internally -- it just keeps track of offsets. It depends on the user
how (s)he accesses the data and what the offsets mean.
It allows for more efficient implementations when we know the size of
events.